data/requirements.txt
requests
python-dotenv
pandas
scipy
torch
onnx
